backend/requirements.txt
fastapi
uvicorn
kafka-python
psycopg2-binary
scikit-learn
numpy
pandas